PYIC-8692: Move cimitStubItemService.ts to common folder
